Handle async builds & compactions on X64 process side.

+++ Source/GenerationsRaytracing.X86/RaytracingRendering.h
@@ -3,7 +3,7 @@
 struct RaytracingRendering
 {
     static inline Hedgehog::Math::CVector s_worldShift;
-    static inline uint64_t s_frame;
+    static inline uint32_t s_frame;
 
     static void init();
     static void postInit();
